+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Match 11 2023 1.19.0
+ * Update modifier option with parent (breaking change)
+
 ## Oct 27th 2021 1.16.0
  * Added cookie support fo http/s fs
  * Added *ssh.ClientConfig option